dependency updates

+++ src/main/java/be/e_contract/ethereum/tool/TransactionTemplate.java
@@ -1,6 +1,6 @@
 /*
  * Ethereum Tool project.
- * Copyright (C) 2018 e-Contract.be BVBA.
+ * Copyright (C) 2018-2020 e-Contract.be BV.
  *
  * This is free software; you can redistribute it and/or modify it
  * under the terms of the GNU Lesser General Public License version
@@ -36,5 +36,5 @@ public class TransactionTemplate {
     public long nonce;
 
     // optional
-    public Byte chainId;
+    public Long chainId;
 }
